Updated version of wp and wc, bug fixses, some major updates to plugins

+++ wp-content/plugins/learnpress/inc/class-lp-hard-cache.php
@@ -25,27 +25,67 @@ class LP_Hard_Cache {
 	protected static $_lock = false;
 
 	/**
+	 * @var bool
+	 */
+	protected static $_hard_lock = false;
+
+	/**
 	 * Init
 	 */
 	public static function init() {
 		$upload_dir       = wp_upload_dir();
 		self::$_root_path = trailingslashit( $upload_dir['basedir'] ) . 'lp-cache';
 
-		@wp_mkdir_p( self::$_root_path );
+		if ( defined( 'LP_HARD_CACHE' ) ) {
+			self::$_lock = self::$_hard_lock = ! LP_HARD_CACHE;
+		} else {
+			self::$_lock = self::$_hard_lock = ! ( 'yes' === LP()->settings()->get( 'enable_hard_cache' ) );
+		}
 
-		self::$_lock = self::get_option_enable();
+		if ( self::is_locked() ) {
+			return;
+		}
+
+		@wp_mkdir_p( self::$_root_path ); // phpcs:ignore
+	}
+
+	/**
+	 * Replace existing cache with new data.
+	 *
+	 * @param string $key
+	 * @param mixed $data
+	 * @param string $group
+	 *
+	 * @return bool
+	 */
+	public static function replace( $key, $data, $group = '' ) {
+		if ( self::is_locked() ) {
+			return false;
+		}
+
+		$file = self::get_file( $key, $group );
+
+		if ( ! file_exists( $file ) ) {
+			return false;
+		}
+
+		return self::write( $key, $data, $group );
 	}
 
 	/**
 	 * Set cache content, replace if existing.
 	 *
 	 * @param string $key
-	 * @param mixed  $data
+	 * @param mixed $data
 	 * @param string $group
 	 *
 	 * @return bool
 	 */
 	public static function set( $key, $data, $group = '' ) {
+		if ( self::is_locked() ) {
+			return false;
+		}
+
 		return self::write( $key, $data, $group, true );
 	}
 
@@ -54,13 +94,11 @@ class LP_Hard_Cache {
 	 *
 	 * @param string $key
 	 * @param string $group
-	 * @param bool   $force | set true if want always cache
 	 *
 	 * @return bool
 	 */
-	public static function get( $key, $group = '', $force = false ) {
-
-		if ( self::is_locked() && ! $force ) {
+	public static function get( $key, $group = '' ) {
+		if ( self::is_locked() ) {
 			return false;
 		}
 
@@ -75,11 +113,14 @@ class LP_Hard_Cache {
 	 *
 	 * @return bool|mixed
 	 */
-	protected static function read( $key, $group = '' ) {
-		$file = self::get_file( $key, $group );
+	public static function read( $key, $group = '' ) {
+		if ( self::is_locked() ) {
+			return false;
+		}
 
+		$file = self::get_file( $key, $group );
 		if ( file_exists( $file ) ) {
-			$f       = @fopen( $file, 'r' );
+			$f       = @fopen( $file, 'r' ); // phpcs:ignore
 			$content = fread( $f, filesize( $file ) );
 			fclose( $f );
 
@@ -97,13 +138,18 @@ class LP_Hard_Cache {
 	 * Write data into file.
 	 *
 	 * @param string $key
-	 * @param mixed  $data
+	 * @param mixed $data
 	 * @param string $group
-	 * @param bool   $overwrite
+	 * @param bool $overwrite
 	 *
 	 * @return bool
 	 */
 	protected static function write( $key, $data, $group = '', $overwrite = false ) {
+
+		if ( self::is_locked() ) {
+			return false;
+		}
+
 		$file = self::get_file( $key, $group );
 
 		/**
@@ -134,10 +180,14 @@ class LP_Hard_Cache {
 	 * @return string
 	 */
 	protected static function get_file( $key, $group = '' ) {
+		if ( self::is_locked() ) {
+			return false;
+		}
+
 		$path = array( self::$_root_path, $group );
 		$path = array_filter( $path );
 		$path = join( '/', $path );
-		@wp_mkdir_p( $path );
+		@wp_mkdir_p( $path ); // phpcs:ignore
 		$file = "{$path}/{$key}.cache";
 
 		return $file;
@@ -152,6 +202,27 @@ class LP_Hard_Cache {
 	}
 
 	/**
+	 * Enable for updating cache temporary even it is locked.
+	 */
+	public static function unlock() {
+		self::$_lock = false;
+	}
+
+	/**
+	 * Disable for updating cache temporary event it is not locked.
+	 */
+	public static function lock() {
+		self::$_lock = true;
+	}
+
+	/**
+	 * Reset lock to default.
+	 */
+	public static function reset_lock() {
+		self::$_lock = self::$_hard_lock;
+	}
+
+	/**
 	 * Check if cache is locked
 	 *
 	 * @return bool
@@ -160,19 +231,16 @@ class LP_Hard_Cache {
 		return self::$_lock;
 	}
 
-	public static function get_option_enable() {
-		return 'yes' === get_option( 'learn_press_enable_hard_cache' );
-	}
-
-	/**
-	 * Clear cache
-	 *
-	 * @param false $group
-	 *
-	 * @return mixed
-	 */
 	public static function flush( $group = false ) {
-		$wp_filesystem = LP_Helper::get_wp_filesystem();
+		if ( ! function_exists( 'WP_Filesystem' ) ) {
+			require_once( ABSPATH . 'wp-admin/includes/file.php' );
+		}
+		WP_Filesystem();
+		global $wp_filesystem;
+
+		if ( ! $wp_filesystem ) {
+			return false;
+		}
 
 		if ( $group ) {
 			$return = $wp_filesystem->rmdir( self::get_path( $group ), true );
